requirements.txt
numpy==1.19.5
tensorflow==2.5.0 
pandas==1.3.0 
transformers==4.8.2 
spacy==3.1.1 
progressbar2==3.53.1 
streamlit==0.84.1 
